requirements.txt
openai
vllm
pandas
pyyaml
python-dotenv
# tensorrt_llm (optional, if using TensorRT-LLM) 